second_patchupsampling_withsem/semencoder/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
matplotlib>=3.5.0
Pillow>=8.3.0
opencv-python>=4.5.0
tensorboard>=2.8.0
argparse
pathlib
typing
json
datetime 